fixed isort issue with circular import generation
