+++ .isoft.cfg
@@ -5,4 +5,3 @@ known_first_party = app,config
 known_third_party = pypika,motor,asyncio_pool,uvloop
 multi_line_output = 0
 sections = FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER
-extend_skip_glob = deepint/util/__init__.py
