+++ .pre-commit-config.yaml
@@ -12,10 +12,10 @@ repos:
     hooks:
     -   id: flake8
 
-#-   repo: https://github.com/pre-commit/mirrors-isort
-#    rev: v4.3.21
-#    hooks:
-#    -   id: isort
+-   repo: https://github.com/pre-commit/mirrors-isort
+    rev: v4.3.21
+    hooks:
+    -   id: isort
 -   repo: https://github.com/asottile/pyupgrade
     rev: v1.17.1
     hooks:
