+++ deepint/core/__init__.py
@@ -7,7 +7,9 @@ from .alert import Alert, AlertType
 from .dashboard import Dashboard
 from .model import Model, ModelMethod, ModelType
 from .organization import Organization
-from .source import DerivedSourceType, ExternalSource, FeatureType, RealTimeSource, AutoUpdatedSource, Source, SourceFeature, SourceType
+from .source import (AutoUpdatedSource, DerivedSourceType, ExternalSource,
+                     FeatureType, RealTimeSource, Source, SourceFeature,
+                     SourceType)
 from .task import Task, TaskStatus
 from .visualization import Visualization
 from .workspace import Workspace
