+++ deepint/util/__init__.py
@@ -5,7 +5,5 @@
 
 
 from .parser import parse_date, parse_url
-from .request import handle_paginated_request, handle_request
-
-# Note: import performed here to avoid circular import error
-from .custom_endpoint import CustomEndpointCall
+from .request import (CustomEndpointCall, handle_paginated_request,
+                      handle_request)
